Added smart pointers, added leaky ReLu, fixed batch averaging bug, fixed activation partial derivative bug

+++ Include/DataManager.h
@@ -8,11 +8,11 @@ public:
 	~DataManager();
 
 	int ReverseInt(int i);
-	std::vector<std::pair<std::vector<float>, int>> LoadImageData(std::string path, std::string labelsPath, int numberOfImages);
+	std::vector<std::pair<std::vector<double>, int>> LoadImageData(std::string path, std::string labelsPath, int numberOfImages);
 	void ShuffleTrainingData();
 
-	std::vector<std::pair<std::vector<float>, int>> trainingData;
-	std::vector<std::pair<std::vector<float>, int>> testingData;
+	std::vector<std::pair<std::vector<double>, int>> trainingData;
+	std::vector<std::pair<std::vector<double>, int>> testingData;
 
 private:
