+++ Include/DisplayManager.h
@@ -9,7 +9,7 @@ public:
 	DisplayManager();
 	std::string UserInputOnTrainingCompleted();
 	void DrawNetworkPredictions(NeuralNetwork& network, DataManager& dataManager);
-	void DrawNumber(const std::pair<std::vector<float>, int>& imageData, const int networkPrediction);
+	void DrawNumber(const std::pair<std::vector<double>, int>& imageData, const int networkPrediction);
 
 	int numImagesToDisplay;
 private:
