+++ Include/LayerResults.h
@@ -5,9 +5,9 @@ class LayerResults
 public:
 	LayerResults() = default;
 	LayerResults(int previousLayerSize, int currentLayerSize);
-	LayerResults(std::vector<std::vector<long float>> weight, std::vector<long float>bias) : mWeightedResults{ weight }, mBiasResults{ bias } {}
-	std::vector<std::vector<long float>> mWeightedResults;
-	std::vector<long float> mBiasResults;
+	LayerResults(std::vector<std::vector<double>> weight, std::vector<double>bias) : mWeightedResults{ weight }, mBiasResults{ bias } {}
+	std::vector<std::vector<double>> mWeightedResults;
+	std::vector<double> mBiasResults;
 
 	LayerResults operator+ (LayerResults obj)
 	{
@@ -37,7 +37,7 @@ public:
 		return { mWeightedResults, mBiasResults };
 	}
 
-	LayerResults operator *(float val)
+	LayerResults operator *(double val)
 	{
 		for (int i = 0; i < mWeightedResults.size(); i++)
 		{
